Update pr.yml
Commenting on PR status for forked repos won't work.
Could create a repo secret and use that to comment but that secret can only live for 90 days and needs to constantly be rotated.

Otherwise we would have to setup a GitHub app and a Key Vault and use that to generate a token.

But the easiest way to deal with this, just comment on PRs that are not forked. 

That means for forked PRs, you have to manually check coverage looking at PR action.